// Angular/src/app/models/movie.ts
export interface Movie {
    userId: string;
    movieId: string;
    movieTitle: string;
    moviePoster: string;
    actors: string;
    director: string;
    genre: string;
    releaseDate: string;
    overview: string;
    comment: string;
    rating: number;
    createdAt: string;
    updatedAt: string;
    time: number;
    }